pyproject.toml
[project]
name = "slothusseeker"
version = "0.1.0"
description = "Discord bot for scraping SimplifyJobs internship listings from GitHub"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]
